Rename some things.
